add: like and tag

+++ post/views.py
@@ -1,9 +1,10 @@
 from django.shortcuts import render
 
-# FBV
+# FBV: 이거 안 씀.
 
 from rest_framework.response import Response
-from .models import Post
+from .models import Post, Like
+from tag.models import Tag
 from rest_framework.decorators import api_view
 from django.views.decorators.csrf import csrf_exempt
 
@@ -25,7 +26,7 @@ def ReadAllPostView(request):
     return Response({"posts":contents})
 
 
-# CBV
+# CBV: 우린 이걸로 함.
 
 from rest_framework.views import APIView
 from rest_framework import status
@@ -39,11 +40,23 @@ class PostListView(APIView):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def post(self, request):
+        author = request.user
         title = request.data.get('title')
         content = request.data.get('content')
+        tag_ids = request.data.get('tags')
+
+        if not author.is_authenticated:
+            return Response({"detail": "Authentication credentials not provided"}, status=status.HTTP_401_UNAUTHORIZED)
+
         if not title or not content:
             return Response({"detail": "[title, content] fields missing."}, status=status.HTTP_400_BAD_REQUEST)
-        post = Post.objects.create(title=title, content=content)
+        
+        for tag_id in tag_ids:
+            if not Tag.objects.filter(id=tag_id).exists():
+                return Response({"detail": "Provided tag not found."}, status=status.HTTP_404_NOT_FOUND)
+
+        post = Post.objects.create(title=title, content=content, author=author)
+        post.tags.set(tag_ids)
         serializer = PostSerializer(post)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
@@ -61,19 +74,65 @@ class PostDetailView(APIView):
             post = Post.objects.get(id=post_id)
         except:
             return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+        
+        if request.user != post.author:
+            return Response({"detail": "Permission denied"}, status=status.HTTP_401_UNAUTHORIZED)
+        
         post.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
     
-    def put(self, request, post_id):
+    # def put(self, request, post_id):
+    #     try:
+    #         post = Post.objects.get(id=post_id)
+    #         # post.title = request.data.get('title')
+    #         # post.content = request.data.get('content')
+    #     except:
+    #         return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+        
+    #     serializer = PostSerializer(post, data=request.data)
+
+    #     if (serializer.is_valid()):
+    #       serializer.save()
+    #       return Response(serializer.data, status=status.HTTP_200_OK)
+
+    def patch(self, request, post_id):
         try:
             post = Post.objects.get(id=post_id)
-            # post.title = request.data.get('title')
-            # post.content = request.data.get('content')
         except:
             return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
-        serializer = PostSerializer(post, data=request.data)
-        if (serializer.is_valid()):
-          serializer.save()
-          return Response(serializer.data, status=status.HTTP_200_OK)
-      
+        
+        if request.user != post.author:
+            return Response({"detail": "Permission denied"}, status=status.HTTP_401_UNAUTHORIZED)
+
+        serializer = PostSerializer(post, data=request.data, partial=True)
+
+        if not serializer.is_valid():
+            return Response({"detail": "data validation error"}, status=status.HTTP_400_BAD_REQUEST)
+        
+        serializer.save()
+        return Response(serializer.data, status=status.HTTP_200_OK)
+    
+
+class LikeView(APIView):
+    def post(self, request, post_id):
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials not provided"}, status=status.HTTP_401_UNAUTHORIZED)
+
+        ### 1 ###
+        try:
+            post = Post.objects.get(id=post_id)
+        except:
+            return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+
+        ### 2 ###
+        like_list = post.like_set.filter(user=request.user)
+
+        ### 3 ###
+        if like_list.count() > 0: # 사실상 1 (좋아요 누른 적 있다면)
+            post.like_set.get(user=request.user).delete() # 좋아요 취소
+        else:
+            Like.objects.create(user=request.user, post=post)
+
+        serializer = PostSerializer(instance=post)
+        return Response(serializer.data, status=status.HTTP_200_OK)
